update code
the cloudinary rename method does not change the stored url, can only change it in the cloudinary database, so I just deleted old and added new with new url. name is the same as public id, so I had to do some weird things so that if the name changed, I also updated the avatar image.